Refactored application controller to include concer for source feature and devise whitelist
