+++ app/controllers/application_controller.rb
@@ -1,16 +1,5 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
-
-  before_filter :configure_permitted_parameters, if: :devise_controller?
-
-  def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:sign_up, keys: [:name])
-    devise_parameter_sanitizer.permit(:account_update, keys: [:name])
-  end
-
-  before_action :set_source
-
-  def set_source
-    session[:source] = params[:q] if params[:q]
-  end
+  include DeviseWhitelist
+  include SetSource
 end
