+++ app/controllers/concerns/set_source.rb
@@ -1,0 +1,11 @@
+module SetSource
+  extend ActiveSupport::Concern
+
+  included do
+    before_filter :set_source
+  end
+
+  def set_source
+    session[:source] = params[:q] if params[:q]
+  end
+end
